feat(audio-backend): replace WebSocket transport with Tauri Channel for audio events
Refactor the native audio backend's frontend communication from a split
local-WebSocket transport (separate event/control sockets) to Tauri's
native `Channel` API for events and `invoke` for commands.

- Events (FFT, position, status) now stream over a `tauri::ipc::Channel`
  registered via new `audio_subscribe_events` command. Works on all
  webview backends (WebView2, WebKitGTK, WKWebView, Android WebView)
  where
  the previous loopback WebSocket could not be consumed.
- Playback commands (play/pause/seek/volume) go over the existing
  `audio_send_msg` invoke — a separate IPC path so controls never queue
  behind high-rate FFT frames.
- Removed `audioWs.ts`, `ws_server.rs`, `audio_get_ws_url(s)` commands,
  and `tokio-tungstenite`/`futures-util` dependencies.
- AutoMix analysis split into `analysis.rs` (signal primitives) and
  `structure.rs` (intro/outro/section detection) for clarity.
- Mixer gains a `fill_scaled` fast path for bulk sample copying with
  gain.
- Added `core-graphics` dependency for macOS window traffic-light
  insets.

+++ src-tauri/src/desktop/window/config.rs
@@ -92,7 +92,7 @@ pub struct WindowConfig {
     pub use_overlay_titlebar: bool,
     /// macOS traffic lights inset (x, y). Only used on macOS.
     #[serde(default)]
-    pub traffic_lights_inset: Option<(f64, f64)>,
+    pub traffic_lights_inset: Option<(f32, f32)>,
     /// Native window effect to apply (e.g. "acrylic"). Platform-specific.
     #[serde(default)]
     pub window_effect: Option<String>,
